@@ -19,7 +19,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.1.1+11
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.5.3
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,31 +36,33 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  localstorage: ^4.0.1+4
-  upgrader: ^10.3.0
-  flutter_launcher_icons: ^0.13.1
+  localstorage: ^5.0.0
+  upgrader: ^11.2.0
+  new_version_plus: ^0.0.9
+  flutter_launcher_icons: ^0.14.1
   change_app_package_name: ^1.1.0
   font_awesome_flutter: ^10.4.0
   flutter_wall_layout: ^2.1.1
-  logger: ^1.3.0
+  logger: ^2.4.0
   dropdown_button2: ^2.1.2
-  device_info_plus: ^9.1.2
+  device_info_plus: ^10.1.2
+  package_info_plus: ^8.0.2
   focus_detector_v2: ^3.0.0
-  syncfusion_flutter_gauges: ^21.2.6
-  mqtt5_client: ^3.4.0
+  syncfusion_flutter_gauges: ^27.1.51
+  mqtt5_client: ^4.5.3
   url_launcher: ^6.1.11
   liquid_progress_indicator_v2: ^0.5.0
-  http: ^0.13.5
+  http: ^1.2.2
   xml: ^6.2.2
   shared_preferences: ^2.1.0
-  intl: ^0.18.0
-  slide_digital_clock: ^1.0.3
+  intl: ^0.19.0
+  one_clock: ^2.0.1
   turkish: ">=0.2.1"
   flutter_keyboard_visibility: ^6.0.0
   flutter_advanced_switch: ^3.0.1
-  flutter_inappwebview: ^5.7.2+3
-  calendar_date_picker2: ^0.5.2
-  flutter_localization: ^0.1.12
+  flutter_inappwebview: ^6.1.4
+  calendar_date_picker2: ^1.1.7
+  flutter_localization: ^0.2.2
   text_scroll: ^0.2.0
   introduction_screen: ^3.1.14
   permission_handler: ^11.3.1
@@ -73,8 +75,8 @@ dependencies:
   hive_flutter: ^1.1.0
   flutter_background_service: ^5.0.10
   flutter_local_notifications: ^17.2.2
-  dropdown_search: ^5.0.6
-  modal_bottom_sheet: ^2.1.2
+  dropdown_search: ^6.0.1
+  modal_bottom_sheet: ^3.0.0
 
 flutter_icons:
   android: "launcher_icon"
@@ -89,7 +91,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
