@@ -2,7 +2,7 @@ import 'package:akilli_anahtar/utils/constants.dart';
 import 'package:flutter/material.dart';
 
 class NullItem extends StatelessWidget {
-  const NullItem({Key? key}) : super(key: key);
+  const NullItem({super.key});
 
   @override
   Widget build(BuildContext context) {
