@@ -3,8 +3,7 @@ import 'package:flutter/material.dart';
 class CustomContainer extends StatelessWidget {
   final Widget body;
   final int flex;
-  const CustomContainer({Key? key, required this.body, this.flex = 5})
-      : super(key: key);
+  const CustomContainer({super.key, required this.body, this.flex = 5});
 
   @override
   Widget build(BuildContext context) {
