@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:akilli_anahtar/entities/box.dart';
 import 'package:akilli_anahtar/models/version_model.dart';
 import 'package:akilli_anahtar/services/api/base_service.dart';
@@ -10,7 +8,7 @@ class BoxService {
   static String url = "$apiUrlOut/Box";
 
   static Future<Box?> get(int id) async {
-    var response = await BaseService.get(url, id);
+    var response = await BaseService.get("$url/get?id=$id");
     if (response.statusCode == 200) {
       var box = Box.fromJson(response.body);
       return box;
@@ -19,17 +17,17 @@ class BoxService {
   }
 
   static Future<List<Box>?> getAll() async {
-    var response = await BaseService.getAll(url);
+    var response = await BaseService.get(
+      "$url/getAll",
+    );
     if (response.statusCode == 200) {
-      var boxList = List<Box>.from(response.body as List<dynamic>)
-          .map((e) => Box.fromJson(json.encode(e)));
-      return boxList.toList();
+      return Box.fromJsonList(response.body);
     }
     return null;
   }
 
   static Future<Box?> add(Box box) async {
-    var response = await BaseService.add(url, box.toJson());
+    var response = await BaseService.add("$url/add", box.toJson());
     if (response.statusCode == 200) {
       var box = Box.fromJson(response.body);
       return box;
@@ -38,7 +36,7 @@ class BoxService {
   }
 
   static Future<Box?> update(Box box) async {
-    var response = await BaseService.update(url, box.toJson());
+    var response = await BaseService.update("$url/update", box.toJson());
     if (response.statusCode == 200) {
       var box = Box.fromJson(response.body);
       return box;
@@ -46,13 +44,12 @@ class BoxService {
     return null;
   }
 
-  static Future<Box?> delete(int id) async {
-    var response = await BaseService.delete(url, id);
+  static Future<bool> delete(int id) async {
+    var response = await BaseService.delete("$url/delete?id=$id");
     if (response.statusCode == 200) {
-      var box = Box.fromJson(response.body);
-      return box;
+      return true;
     }
-    return null;
+    return false;
   }
 
   static Future<VersionModel> checkNewVersion() async {
