@@ -2,6 +2,7 @@ import 'package:akilli_anahtar/entities/city.dart';
 import 'package:akilli_anahtar/entities/district.dart';
 import 'package:akilli_anahtar/entities/operation_claim.dart';
 import 'package:akilli_anahtar/entities/organisation.dart';
+import 'package:akilli_anahtar/entities/user_organisation.dart';
 import 'package:akilli_anahtar/services/api/base_service.dart';
 import 'package:akilli_anahtar/utils/constants.dart';
 
@@ -9,58 +10,93 @@ class HomeService {
   static String url = "$apiUrlOut/Home";
 
   static Future<City?> getCity(int id) async {
-    return BaseService.get2(
-      "$url/getCity?id=",
-      (json) => City.fromJson(json),
+    var response = await BaseService.get(
+      "$url/getCity?id=$id",
     );
+    if (response.statusCode == 200) {
+      return City.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<List<City>?> getAllCity(int id) async {
-    return BaseService.get2(
+  static Future<List<City>?> getAllCity() async {
+    var response = await BaseService.get(
       "$url/getAllCity",
-      (json) => City.fromJsonList(json),
     );
+    if (response.statusCode == 200) {
+      return City.fromJsonList(response.body);
+    }
+    return null;
   }
 
   static Future<District?> getDistrict(int id) async {
-    return BaseService.get2(
-      "$url/getDistrict?id=",
-      (json) => District.fromJson(json),
+    var response = await BaseService.get(
+      "$url/getDistrict?id=$id",
     );
+    if (response.statusCode == 200) {
+      return District.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<List<District>?> getAllDistrict(int id) async {
-    return BaseService.get2(
+  static Future<List<District>?> getAllDistrict() async {
+    var response = await BaseService.get(
       "$url/getAllDistrict",
-      (json) => District.fromJsonList(json),
     );
+    if (response.statusCode == 200) {
+      return District.fromJsonList(response.body);
+    }
+    return null;
   }
 
   static Future<OperationClaim?> getOperationClaim(int id) async {
-    return BaseService.get2(
-      "$url/getOperationClaim?id=",
-      (json) => OperationClaim.fromJson(json),
+    var response = await BaseService.get(
+      "$url/getOperationClaim?id=$id",
     );
+    if (response.statusCode == 200) {
+      return OperationClaim.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<List<OperationClaim>?> getAllOperationClaim(int id) async {
-    return BaseService.get2(
+  static Future<List<OperationClaim>?> getAllOperationClaim() async {
+    var response = await BaseService.get(
       "$url/getAllOperationClaim",
-      (json) => OperationClaim.fromJsonList(json),
     );
+    if (response.statusCode == 200) {
+      return OperationClaim.fromJsonList(response.body);
+    }
+    return null;
   }
 
   static Future<Organisation?> getOrganisation(int id) async {
-    return BaseService.get2(
-      "$url/getOrganisation?id=",
-      (json) => Organisation.fromJson(json),
+    var response = await BaseService.get(
+      "$url/getOrganisation?id=$id",
     );
+    if (response.statusCode == 200) {
+      return Organisation.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<List<Organisation>?> getAllOrganisation(int id) async {
-    return BaseService.get2(
+  static Future<List<Organisation>?> getAllOrganisation() async {
+    var response = await BaseService.get(
       "$url/getAllOrganisation",
-      (json) => Organisation.fromJsonList(json),
     );
+    if (response.statusCode == 200) {
+      return Organisation.fromJsonList(response.body);
+    }
+    return null;
+  }
+
+  static Future<List<UserOrganisation>?> getUserOrganisationsByOrganisationId(
+      int organisationId) async {
+    var response = await BaseService.get(
+      "$url/getUserOrganisationsByOrganisationId?organisationId=$organisationId",
+    );
+    if (response.statusCode == 200) {
+      return UserOrganisation.fromJsonList(response.body);
+    }
+    return null;
   }
 }
