@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-import 'package:akilli_anahtar/entities/operation_claim.dart';
 import 'package:akilli_anahtar/entities/user.dart';
 import 'package:akilli_anahtar/entities/user_device.dart';
+import 'package:akilli_anahtar/entities/user_operation_claim.dart';
 import 'package:akilli_anahtar/entities/user_organisation.dart';
 import 'package:akilli_anahtar/services/api/base_service.dart';
 import 'package:akilli_anahtar/utils/constants.dart';
@@ -11,40 +11,66 @@ class UserService {
   static String url = "$apiUrlOut/User";
 
   static Future<User?> get(int id) async {
-    return BaseService.get2(
+    var response = await BaseService.get(
       "/get?id=",
-      (json) => User.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return User.fromJson(response.body);
+    }
+    return null;
   }
 
   static Future<User?> getbyUserName(String userName) async {
-    return BaseService.get2(
+    var response = await BaseService.get(
       "$url/getByUserName?userName=$userName",
-      (json) => User.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return User.fromJson(response.body);
+    }
+    return null;
   }
 
   static Future<List<User>?> getAll() async {
-    return BaseService.get2(
+    var response = await BaseService.get(
       "$url/getAll",
-      (json) => User.fromJsonList(json),
     );
+    if (response.statusCode == 200) {
+      return User.fromJsonList(response.body);
+    }
+    return null;
   }
 
   static Future<User?> register(User user) async {
-    return BaseService.add2(
+    var response = await BaseService.add(
       "$url/register",
       user.toJson(),
-      (json) => User.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return User.fromJson(response.body);
+    }
+    return null;
+  }
+
+  static Future<User?> saveAs(User user) async {
+    var response = await BaseService.add(
+      "$url/saveAs",
+      user.toJson(),
+    );
+    if (response.statusCode == 200) {
+      return User.fromJson(response.body);
+    }
+    return null;
   }
 
   static Future<User?> update(User user) async {
-    return BaseService.add2(
-      "$url/register",
+    var response = await BaseService.update(
+      "$url/update",
       user.toJson(),
-      (json) => User.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return User.fromJson(response.body);
+    }
+    return null;
   }
 
   static Future<bool> passwordUpdate(int id, String password) async {
@@ -52,53 +78,106 @@ class UserService {
       'id': id,
       'password': password,
     };
-    return BaseService.update2(
-        url, json.encode(requestBody), (json) => json.contains("true"));
-  }
-
-  static Future<User> delete(id) async {
-    return BaseService.delete2(
-      "$url/delete?id=$id",
-      (json) => User.fromJson(json),
+    var response = await BaseService.update(
+      url,
+      json.encode(requestBody),
     );
+    if (response.statusCode == 200) {
+      return true;
+    }
+    return false;
   }
 
-  static Future<OperationClaim?> addClaim(int userId, int claimId) async {
+  static Future<bool> delete(int id) async {
+    var response = await BaseService.delete(
+      "$url/delete?id=$id",
+    );
+    if (response.statusCode == 200) {
+      return true;
+    }
+    return false;
+  }
+
+  static Future<List<UserOperationClaim>?> getUserClaims(int userId) async {
+    var response = await BaseService.get(
+      "$url/getUserClaims?userId=$userId",
+    );
+    if (response.statusCode == 200) {
+      return UserOperationClaim.fromJsonList(response.body);
+    }
+    return null;
+  }
+
+  static Future<UserOperationClaim?> addUserClaim(
+      int userId, int claimId) async {
     var requestBody = {
       'userId': userId,
       'claimId': claimId,
     };
-    return BaseService.add2(
-      "$url/addClaim?userId=$userId&claimId=$claimId",
+    var response = await BaseService.add(
+      "$url/addUserClaim?userId=$userId&claimId=$claimId",
       json.encode(requestBody),
-      (json) => OperationClaim.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return UserOperationClaim.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<OperationClaim> deleteClaim(int userClaimId) async {
-    return BaseService.delete2(
-      "$url/deleteClaim?userClaimId=$userClaimId",
-      (json) => OperationClaim.fromJson(json),
+  static Future<bool> deleteUserClaim(int userClaimId) async {
+    var response = await BaseService.delete(
+      "$url/deleteUserClaim?userClaimId=$userClaimId",
     );
+    if (response.statusCode == 200) {
+      return true;
+    }
+    return false;
   }
 
-  static Future<UserDevice?> addDevice(int userId, int deviceId) async {
+  static Future<List<UserDevice>?> getUserDevices(int userId) async {
+    var response = await BaseService.get(
+      "$url/getUserDevices?userId=$userId",
+    );
+    if (response.statusCode == 200) {
+      return UserDevice.fromJsonList(response.body);
+    }
+    return null;
+  }
+
+  static Future<UserDevice?> addUserDevice(int userId, int deviceId) async {
     var requestBody = {
       'userId': userId,
       'deviceId': deviceId,
     };
-    return BaseService.add2(
-      "$url/addDevice?userId=$userId&deviceId=$deviceId",
+    var response = await BaseService.add(
+      "$url/addUserDevice?userId=$userId&deviceId=$deviceId",
       json.encode(requestBody),
-      (json) => UserDevice.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return UserDevice.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<UserDevice> deleteUserDevice(int userDeviceId) async {
-    return BaseService.delete2(
+  static Future<bool> deleteUserDevice(int userDeviceId) async {
+    var response = await BaseService.delete(
       "$url/deleteUserDevice?userDeviceId=$userDeviceId",
-      (json) => UserDevice.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return true;
+    }
+    return false;
+  }
+
+  static Future<List<UserOrganisation>?> getUserOrganisations(
+      int userId) async {
+    var response = await BaseService.get(
+      "$url/getUserOrganisations?userId=$userId",
+    );
+    if (response.statusCode == 200) {
+      return UserOrganisation.fromJsonList(response.body);
+    }
+    return null;
   }
 
   static Future<UserOrganisation?> addUserOrganisation(
@@ -107,18 +186,23 @@ class UserService {
       'userId': userId,
       'organisationId': organisationId,
     };
-    return BaseService.add2(
+    var response = await BaseService.add(
       "$url/addUserOrganisation?userId=$userId&organisationId=$organisationId",
       json.encode(requestBody),
-      (json) => UserOrganisation.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return UserOrganisation.fromJson(response.body);
+    }
+    return null;
   }
 
-  static Future<UserOrganisation> deleteOrganisation(
-      int userOrganisationId) async {
-    return BaseService.delete2(
+  static Future<bool> deleteUserOrganisation(int userOrganisationId) async {
+    var response = await BaseService.delete(
       "$url/deleteUserOrganisation?userOrganisationId=$userOrganisationId",
-      (json) => UserOrganisation.fromJson(json),
     );
+    if (response.statusCode == 200) {
+      return true;
+    }
+    return false;
   }
 }
